Data_Analytics/aviation_weather_etl/src/config.py
from dotenv import load_dotenv
import os


load_dotenv()

STATION = os.getenv("STATION", "KNHK")
HOURS_BACK = int(os.getenv("HOURS_BACK", "24"))
DB_URL = os.getenv("DB_URL", "sqlite:///db/aviation.db")
