Data_Analytics/aviation_weather_etl/requirements.txt
pandas
requests
python-dotenv
sqlalchemy
matplotlib
pytest
